Change double to int

+++ Environment_oM/Properties/SpaceAnalyticalProperties.cs
@@ -36,8 +36,8 @@ namespace BH.oM.Environment.Properties
         public double DaylightFactor { get; set; } = 0;
         public double FacadeLength { get; set; } = 0;
         public double FixedConvectionCoefficient { get; set; } = 0;
-        public double SizeCooling { get; set; } = 0;
-        public double SizeHeating { get; set; } = 0;
+        public int SizeCooling { get; set; } = 0;
+        public int SizeHeating { get; set; } = 0;
         public double RadiantProportion { get; set; } = 0;
     }
 }
